PIM-6442: Move reference data to `Value` namespace Move from Pim\Component\ReferenceData\ProductValue to Pim\Component\ReferenceData\Value and Pim\Component\ReferenceData\Factory\ProductValue to Pim\Component\ReferenceData\Factory\Value
