+++ src/Pim/Bundle/ReferenceDataBundle/Resources/config/product_values.yml
@@ -1,3 +1,3 @@
 parameters:
-    pim_reference_data.value.reference_data.class: Pim\Component\ReferenceData\ProductValue\ReferenceDataValue
-    pim_reference_data.value.reference_data_collection.class: Pim\Component\ReferenceData\ProductValue\ReferenceDataCollectionValue
+    pim_reference_data.value.reference_data.class: Pim\Component\ReferenceData\Value\ReferenceDataValue
+    pim_reference_data.value.reference_data_collection.class: Pim\Component\ReferenceData\Value\ReferenceDataCollectionValue
