+++ src/Pim/Component/ReferenceData/Value/ReferenceDataCollectionValue.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\ReferenceData\ProductValue;
+namespace Pim\Component\ReferenceData\Value;
 
 use Pim\Component\Catalog\Model\AbstractValue;
 use Pim\Component\Catalog\Model\AttributeInterface;
