+++ src/Pim/Component/ReferenceData/Value/ReferenceDataValueInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\ReferenceData\ProductValue;
+namespace Pim\Component\ReferenceData\Value;
 
 use Pim\Component\Catalog\Model\ValueInterface;
 use Pim\Component\ReferenceData\Model\ReferenceDataInterface;
@@ -12,15 +12,10 @@ use Pim\Component\ReferenceData\Model\ReferenceDataInterface;
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface ReferenceDataCollectionValueInterface extends ValueInterface
+interface ReferenceDataValueInterface extends ValueInterface
 {
     /**
-     * @return ReferenceDataInterface[]
+     * @return ReferenceDataInterface
      */
     public function getData();
-
-    /**
-     * @return array
-     */
-    public function getReferenceDataCodes();
 }
