+++ src/Pim/Component/ReferenceData/spec/Normalizer/Indexing/Product/ReferenceDataCollectionNormalizerSpec.php
@@ -7,7 +7,7 @@ use Pim\Component\Catalog\AttributeTypes;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\ValueInterface;
 use Pim\Component\ReferenceData\Normalizer\Indexing\Product\ReferenceDataCollectionNormalizer;
-use Pim\Component\ReferenceData\ProductValue\ReferenceDataCollectionValue;
+use Pim\Component\ReferenceData\Value\ReferenceDataCollectionValue;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 class ReferenceDataCollectionNormalizerSpec extends ObjectBehavior
